edgesolution/modules/blobmodule/requirements.txt
azure-iot-device~=2.7.0
azure-storage-blob==12.15.0
